Comet Image Processing
Comet image processing built on Numpy and Scipy libraries for processing
noise and removing the bulk coma to reveal morphological and surface
features such as localized outgasing jets and expanding shells of dust
and gas. Also developed a specialized processing routine for rectangular
to polar unwrapping of images and subsequent Gaussian fitting of
cometary jets in order to analyze the position angle of jets over time
and calculate spin state of comet nuclei.